Moved classes in different directories
- The src/App and src/Utils directories are removed.
- The src/Ui and src/Str directories are created.
