+++ src/Plugin/Bootstrap.php
@@ -10,7 +10,7 @@
  * @link https://github.com/jaxon-php/jaxon-core
  */
 
-namespace Jaxon\App;
+namespace Jaxon\Plugin;
 
 use Jaxon\Config\Config;
 
@@ -139,9 +139,11 @@ class Bootstrap
     /**
      * Wraps the module/package/bundle setup method.
      *
+     * @param boolean        $bSendResponse     Send back a response to the browser after processing the request
+     *
      * @return void
      */
-    public function bootstrap()
+    public function bootstrap($bSendResponse = true)
     {
         $jaxon = jaxon();
         $di = $jaxon->di();
@@ -200,6 +202,9 @@ class Bootstrap
         {
             $jaxon->setOption('core.request.uri', $this->sUri);
         }
+
+        // Whether to send back a response to the browser after processing the request.
+        $jaxon->setOption('core.response.send', $bSendResponse);
 
         // Event after setting up the module
         $app->triggerEvent('post.setup');
